fix: Move out the OIDC tests
